requirements.txt
matplotlib==3.3.0
numpy==1.23.5
pandas==1.0.5
Pillow==9.5.0
scipy==1.10.1
seaborn==0.10.1
tensorflow==2.10.1
